image/binary/geoddist: update geodesic distance transform algorithms to use ChamferMask classes (instead of deprecated chamfer weights)

+++ src/main/java/net/sci/image/binary/distmap/ChamferDistanceTransform3DUInt16.java
@@ -166,7 +166,7 @@ public class ChamferDistanceTransform3DUInt16 extends AlgoStub implements Chamfe
                 for (int x = 0; x < sizeX; x++)
                 {
                     boolean inside = array.getBoolean(x, y, z);
-                    result.setInt(x, y, z, inside ? Integer.MAX_VALUE : 0);
+                    result.setInt(x, y, z, inside ? UInt16.MAX_VALUE : 0);
                 }
             }
         }
